+++ R/analyseDMRs.R
@@ -353,3 +353,54 @@ writeDMRsToBed <- function(dataTable, folder, FDRthres = 0.05) {
 
   invisible(dataTable)
 }
+
+#' Take the top most DMRs per contrast, based on those with the largest value of the selected metric
+#'
+#' @param DMRs  A data frame containing the output of calculateDMRs, potentially with multiple contrasts
+#' @param n How many DMRs to take of each contrast (if that many exists)
+#' @param FDRthres Threshold on the adjusted p values
+#' @param metric Which metric to use to select the top DMRs. Options are deltaBeta, log2FC, adjPval, CpG_density, position (using seqnames and start columns) or any other column in the data frame. 
+#' If `adjPval` or `position` are used, then the window with the smallest value will be chosen, otherwise the largest value will be used.
+#' @param makePositive Whether to reverse the contrast when the window is hypomethylated in the contrast.
+#' @return A data frame with the DMRs with the largest value of the selected metrics
+#' @examples
+#' # calculate some DMRs
+#' DMRs <- exampleTumourNormal %>% calculateDMRs(variable = "type", contrasts = "all", keepContrastMeans = FALSE)
+#' # Find the DMRs with the largest deltaBeta between each comparison:
+#' DMRs %>% sliceDMRs(n = 1)
+#' # Or the windows with the largest log2FC: 
+#' DMRs %>% sliceDMRs(n = 1, metric = log2FC)
+#' # Or the windows with the largest CpG_density: 
+#' DMRs %>% sliceDMRs(n = 1, metric = CpG_density)
+#' # If adjPval is used, then the smallest value is chosen instead:
+#' DMRs %>% sliceDMRs(n = 1, metric = CpG_density)
+#' # If position is used, then the windows are sorted by genomic position:
+#' DMRs %>% sliceDMRs(n = 1, metric = position)
+#' @export
+sliceDMRs <- function(DMRs, n = 1, metric = deltaBeta, makePositive = TRUE, FDRthres = 0.05) {
+  
+  positiveDMRs <- DMRs %>% 
+    pivotDMRsLonger(makePositive = makePositive, FDRthres = FDRthres) %>% 
+    dplyr::relocate(tidyselect::matches("means$"), .after = tidyselect::last_col()) %>%
+    dplyr::group_by(group1, group2)
+
+  if(deparse(substitute(metric)) %in% c("position")){
+    out <- positiveDMRs %>% 
+      dplyr::arrange(seqnames, start) %>% 
+      dplyr::slice(1:(!!n)) %>%
+      dplyr::ungroup()
+  } else if (deparse(substitute(metric)) %in% c("adjPval")){
+    message(glue::glue("Choosing the windows with the smallest value of {rlang::ensym(metric)}"))
+    out <- positiveDMRs %>% 
+      dplyr::arrange({{metric}}) %>% 
+      dplyr::slice(1:(!!n)) %>%
+      dplyr::ungroup()
+  } else {
+    out <- positiveDMRs %>% 
+      dplyr::arrange(dplyr::desc({{metric}})) %>% 
+      dplyr::slice(1:(!!n)) %>%
+      dplyr::ungroup()
+  }
+
+  return(out)
+}