+++ R/utils.R
@@ -360,4 +360,4 @@ asValidGranges <- function(object){
 #' This function captures a printed form of the object x, for use in error messages
 #' Taken from https://stackoverflow.com/a/26083626 by Richie Cotton
 #' @param x An object to capture
-print_and_capture <- function(x) {paste(capture.output(print(x)), collapse = "\n") }
+print_and_capture <- function(x) {paste(utils::capture.output(print(x)), collapse = "\n") }